refactor: rename base component
