+++ src/app/js/base/component.js
