+++ src/components/spoiler-item/spoiler-item.js
@@ -1,4 +1,4 @@
-import Component from '../../app/js/base/Component';
+import Component from '../../app/js/base/component';
 
 class SpoilerItem extends Component {
     bodyElement;
